feat: enhance DrawView component with erase prop and refactor VideoView component
