+++ components/ui/MediaView/VideoView.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from "react";
+import React, { useEffect, useState } from "react";
 import {
   Canvas,
   CanvasProps,
@@ -12,13 +12,18 @@ import {
   useVideo,
 } from "@shopify/react-native-skia";
 import { Dimensions, Pressable, useWindowDimensions, View } from "react-native";
-import { useSharedValue } from "react-native-reanimated";
+import {
+  runOnJS,
+  useDerivedValue,
+  useSharedValue,
+} from "react-native-reanimated";
 import useGlobalStore from "~/store/globalStore";
 import { Audio } from "expo-av";
 import { GestureHandlerRootView } from "react-native-gesture-handler";
 import DrawView from "../DrawView";
 import RenderStickers from "../Stickers/RenderStickers";
 import { GestureHandler } from "../GestureHandler";
+import IconButton from "../IconButton";
 
 type VideoViewProps = Omit<CanvasProps, "children"> & {};
 
@@ -28,34 +33,86 @@ export default function VideoViewComponent(props: VideoViewProps) {
   const paused = useSharedValue(false);
   const { width, height } = Dimensions.get("screen");
 
-  const { video, filter, stickers } = useGlobalStore();
-  const { currentFrame, rotation, size } = useVideo(video, {
+  const { setCameraMode, setIsRecording, video, setVideo, filter, stickers } =
+    useGlobalStore();
+  const { currentFrame, rotation, currentTime } = useVideo(video, {
     paused,
   });
 
   const currentPath = useSharedValue(Skia.Path.Make());
+  const [isPlaying, setIsPlaying] = useState(true);
+  const [muted, setMuted] = useState<boolean>(true);
+  const [curretVideoTime, setCurretVideoTime] = useState<number>(
+    currentTime.value
+  );
+
+  useDerivedValue(() => {
+    if (paused.value) {
+      runOnJS(setCurretVideoTime)(currentTime.value);
+    }
+  }, [currentTime, paused]);
 
   useEffect(() => {
     async function playSound() {
-      console.log("Loading Sound");
       const { sound } = await Audio.Sound.createAsync({ uri: video });
-
-      console.log("Playing Sound");
-      await sound.playAsync();
+      await sound.setIsMutedAsync(muted);
+      if (isPlaying) {
+        await sound.playAsync();
+      } else {
+        await sound.pauseAsync();
+      }
     }
 
     playSound();
-  }, []);
+  }, [muted, isPlaying]);
 
   const src = rect(0, 0, width, height);
   const dst = rect(0, 0, width, height);
   const transform = fitbox("cover", src, dst, rotation);
 
+  const resetAndClose = () => {
+    setVideo("");
+    setCameraMode("picture");
+    setIsRecording(false);
+  };
+
   return (
-    <Pressable
-      style={{ flex: 1 }}
-      onPress={() => (paused.value = !paused.value)}
-    >
+    <View className="flex-1">
+      <View
+        className="mt-28"
+        style={{
+          position: "absolute",
+          right: 6,
+          zIndex: 2,
+          gap: 16,
+        }}
+      >
+        <IconButton
+          onPress={() => {
+            resetAndClose();
+          }}
+          iosName={"xmark"}
+        />
+
+        <IconButton
+          onPress={async () => {
+            setMuted(!muted);
+          }}
+          iosName={!muted ? "speaker.wave.2" : "speaker.slash"}
+        />
+        <IconButton
+          iosName={isPlaying ? "pause" : "play"}
+          onPress={() => {
+            if (isPlaying) {
+              paused.value = !paused.value;
+            } else {
+              paused.value = !paused.value;
+            }
+            setIsPlaying(!isPlaying);
+          }}
+        />
+      </View>
+
       <GestureHandlerRootView style={{ flex: 1, position: "relative" }}>
         <DrawView currentPath={currentPath}>
           <View style={{ flex: 1 }}>
@@ -97,6 +154,6 @@ export default function VideoViewComponent(props: VideoViewProps) {
           </View>
         </DrawView>
       </GestureHandlerRootView>
-    </Pressable>
+    </View>
   );
 }
